# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NPTC Catalogue Maintenance Platform",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/catalogue/entries": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "One page of published catalogue entries",
        "description": "Keyset paging on `business_key`, ascending. Pass the response's\n`next_cursor` back as `after` for the following page; a `null`\n`next_cursor` means there is no following page.\n\nThere is no `offset` and no total count, on purpose (ADR-0024): an\noffset re-reads and re-skips every earlier row on every page, and drops\nor repeats rows outright when a concurrent insert shifts the window\nmid-scan.\n\n`filter.*` parameters are accepted here and behave exactly as they do on\n`/catalogue/search`. Facets are **not** returned: computing counts on\nevery page of a browse costs something no caller has asked for, and\n`GET /catalogue/search` is where the facet list with counts lives\n(ADR-0032).",
        "operationId": "list_entries_api_v1_catalogue_entries_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Maximum entries in this page.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum entries in this page."
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^NPTC-([0-9]{6,})$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The `next_cursor` from the previous page. Pass it back unmodified, and do not construct one.",
              "title": "After"
            },
            "description": "The `next_cursor` from the previous page. Pass it back unmodified, and do not construct one."
          },
          {
            "name": "filter.{property_key}",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Filter by a facet. The parameter name is the facet's `key` prefixed with `filter.` - `?filter.discipline=chemistry`. Repeat the parameter to select several values of one facet; they are OR-ed. Filters on different facets are AND-ed, so adding one always narrows the result. The facets available are not fixed: they are every property an administrator has marked filterable, plus the entry status, and `GET /catalogue/search` returns the current list with counts. An operator other than the default `equals` is named after the key, separated by `:` - `?filter.assay_name:prefix=glu`, or `?filter.volume_ml:range=1..5`. Which operators a facet accepts follows from the property's datatype; one it does not accept is a 422, never a silently ignored parameter. At most 50 distinct values are accepted in one facet's selection (repeated parameter or `:in` list alike); more than that is also a 422. NOTE for generated clients: `{property_key}` above is a placeholder, not a literal parameter name - OpenAPI has no syntax for a templated parameter name, so a generated client typically renders one field named literally `filter.{property_key}`. Sending that literal string is a 422 (`{property_key}` is not a filter this endpoint offers); a real filter parameter's name is built by hand, substituting an actual facet key (see ADR-0032)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryPage"
                }
              }
            }
          },
          "401": {
            "description": "A credential was presented and could not be verified. Sending no credential at all is not an error on these endpoints.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A query or path parameter was unprocessable - a business key that is not `NPTC-nnnnnn`, a blank search query, a cursor this API did not issue (including one issued for a different `q` or a different filter set), a `limit` outside its range, or a `filter.*` parameter naming a facet this endpoint does not offer, an operator the facet does not support, or a value the property cannot hold. A filter is never silently ignored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/catalogue/code/{system_token}/{code}": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "One published catalogue entry, resolved by an exact code (FR-17)",
        "description": "Resolves the same entry `GET /catalogue/entries/{business_key}` and\n`GET /catalogue/lookup?system=...&code=...` would for the same code -\nall three serve byte-identical bodies for one entry (FR-17's\nunambiguous-lookup acceptance criterion).\n\n`system_token` is a short alias registered in\n`nptc.catalogue.code_systems` - `sct` for `http://snomed.info/sct`\ntoday. `code` is never validated against a code shape here: an\nunrecognised code and a malformed one both resolve to nothing and get\nthe identical 404 (see `PUBLIC_CODE_LOOKUP_ERROR_RESPONSES`).",
        "operationId": "read_entry_by_code_api_v1_catalogue_code__system_token___code__get",
        "parameters": [
          {
            "name": "system_token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9-]{0,31}$",
              "description": "A short alias for a code system's URI, e.g. `sct` for `http://snomed.info/sct` (FR-17). An unregistered but well-formed token is a 404, not a 422 - see `docs/architecture/public-api.md`.",
              "examples": [
                "sct"
              ],
              "title": "System Token"
            },
            "description": "A short alias for a code system's URI, e.g. `sct` for `http://snomed.info/sct` (FR-17). An unregistered but well-formed token is a 404, not a 422 - see `docs/architecture/public-api.md`."
          },
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The exact code to resolve.",
              "examples": [
                "49466006"
              ],
              "title": "Code"
            },
            "description": "The exact code to resolve."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryDetail"
                }
              }
            }
          },
          "401": {
            "description": "A credential was presented and could not be verified. Sending no credential at all is not an error on these endpoints.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No published entry matches this system and code - because the system_token (or, on `/lookup`, the system URI) is not registered, or because no active or retired binding for this code names a published entry. Both causes return the identical fixed sentence, which names each registered system as both its token and its URI, so a caller cannot use response text to tell them apart.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A query or path parameter was unprocessable - a business key that is not `NPTC-nnnnnn`, a blank search query, a cursor this API did not issue (including one issued for a different `q` or a different filter set), a `limit` outside its range, or a `filter.*` parameter naming a facet this endpoint does not offer, an operator the facet does not support, or a value the property cannot hold. A filter is never silently ignored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntryPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntrySummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "EntryPage",
        "description": "One page of `EntrySummary` rows, keyset-paginated on `business_key`.\n\nServed by both `catalogue.py`'s public `GET /catalogue/entries`\n(`PUBLIC_STATUSES` only) and `catalogue_admin.py`'s\n`GET /catalogue/admin/entries` (any status) - one shape, the\nsame reason `EntryDetail` is shared rather than duplicated. `next_cursor`\nis `null` on the last page - which is the *only* reliable signal that\npaging is finished. A client must not infer the end from a short page: a\npage can be short and still have a successor."
      },
      "EntrySummary": {
        "properties": {
          "business_key": {
            "type": "string",
            "title": "Business Key"
          },
          "preferred_term": {
            "type": "string",
            "title": "Preferred Term"
          },
          "length": {
            "type": "integer",
            "title": "Length"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "specimen_unconstrained": {
            "type": "boolean",
            "title": "Specimen Unconstrained"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "has_open_finding": {
            "type": "boolean",
            "title": "Has Open Finding"
          },
          "label_provenance": {
            "additionalProperties": {
              "$ref": "#/components/schemas/LabelProvenance"
            },
            "type": "object",
            "title": "Label Provenance"
          }
        },
        "type": "object",
        "required": [
          "business_key",
          "preferred_term",
          "length",
          "status",
          "specimen_unconstrained",
          "updated_at",
          "has_open_finding",
          "label_provenance"
        ],
        "title": "EntrySummary",
        "description": "An entry as it appears in a list or a search result.\n\n`length` is FR-85's published figure - the character count of the\ncatalogue's own preferred term, computed by `CatalogueEntry.length` and\nnever stored, so it cannot drift from the term it describes."
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Every refusal this API makes has this shape - one sentence saying\nwhat to do next, and deliberately nothing else. It never names a role,\na permission or an internal identifier (FR-44, NFR-04)."
      },
...
    }
  }
}